LayoutTests/inspector/page/overrideSetting-AdClickAttributionDebugModeEnabled-expected.txt

CONSOLE MESSAGE: [Ad Click Attribution] Turned Debug Mode on.
Tests for the Page.overrideSetting command.


== Running test suite: Page.overrideSetting
-- Running test case: Page.overrideSetting.AdClickAttributionDebugModeEnabled
Enabling ad click attribution debug mode...
PASS: Should broadcast at least one console message.
